Lint improvements
Adopts the following lints as a baseline (with a few exceptions):

    #![warn(
        clippy::mod_module_files,
        clippy::unwrap_used,
        missing_docs,
        rust_2018_idioms,
        unused_lifetimes,
        unused_qualifications
    )]

+++ base32ct/src/alphabet.rs
@@ -28,7 +28,7 @@ pub trait Alphabet: 'static + Copy + Debug + Eq + Send + Sized + Sync {
             // Compute exclusive range from inclusive one
             let start = *range.start() as i16 - 1;
             let end = *range.end() as i16 + 1;
-            ret += (((start - src as i16) & (src as i16 - end)) >> 8) & (src as i16 + *offset);
+            ret += (((start - src) & (src - end)) >> 8) & (src + *offset);
         }
 
         ret
